Created initial JavaScript logic for reading the game board and determining a winner.
The game board is represented by a single array of length 9. An "x", "o" or "" is stored in each element to represent the current state of play. 8 variables were created to store the array elements representing a single winning combination. A "win check" function is used to iterate through each of the 8 winning combinations to determine if any of the arrays contain all x's or o's. If a variable contains an array of all x's or o's, the function currently returns 'player x (or o), you win!'.