Created a basic game board ui with event handlers to manage populating a square with an 'x' or 'o'.
Updated the index.html file with a 3x3 responsive grid using bootstrap. Added event handlers to a new events.js file stored in a new "game" folder where the ui and api js docs will also go. The index.js file now has a populateSquare function that populates the selected board square with an 'x' or 'o' with each consecutive turn. An event handler was added to this function to disable the click event for the selected square to ensure a player cannot update it once it is taken. Simple css styles were added to index.scss to center the 'x' and 'o' text within the div and change the background color to pink for now.